aquarium/requirements.txt
# Anti-Gravity Virtual Aquarium Engine — Dependencies
# Author: Sharjeel Safdar | CG Final Project (All-Time Best Edition)
# Install with:  pip install -r requirements.txt

PyOpenGL>=3.1.7
PyOpenGL_accelerate>=3.1.7
pygame>=2.5.0
numpy>=1.24.0
Pillow>=10.0.0
